+++ sfall/IniReader.h
@@ -20,6 +20,7 @@
 
 namespace sfall
 {
+class Config;
 
 class IniReader {
 public:
@@ -30,6 +31,10 @@ public:
 	static const char* GetConfigFile();
 	static void SetDefaultConfigFile();
 	static void SetConfigFile(const char* iniFile);
+
+	// Gets a Config from an INI file at given path, relative to game root folder.
+	// Config is loaded once per given path when requested and only unloaded on game reset (returning to main menu).
+	static Config* GetIniConfig(const char* iniFile);
 
 	// Gets the integer value from the default config (i.e. ddraw.ini)
 	static int GetIntDefaultConfig(const char* section, const char* setting, int defaultValue);
