+++ sfall/Modules/Scripting/Arrays.h
@@ -203,11 +203,8 @@ ScriptValue GetArrayKey(DWORD id, int index);
 // get array element by index (list) or key (map)
 ScriptValue GetArray(DWORD id, const ScriptValue& key);
 
-// set array element by index or key (with checking the existence of the array ID)
-void SetArray(DWORD id, const ScriptValue& key, const ScriptValue& val, bool allowUnset);
-
 // set array element by index or key
-void setArray(DWORD id, const ScriptValue& key, const ScriptValue& val, bool allowUnset);
+void SetArray(DWORD id, const ScriptValue& key, const ScriptValue& val, bool allowUnset);
 
 // number of elements in list or pairs in map
 int LenArray(DWORD id);
