+++ sfall/Modules/Scripting/Opcodes.cpp
@@ -26,6 +26,7 @@
 #include "Handlers\Core.h"
 #include "Handlers\FileSystem.h"
 #include "Handlers\Graphics.h"
+#include "Handlers\IniFiles.h"
 #include "Handlers\Interface.h"
 #include "Handlers\Inventory.h"
 #include "Handlers\Math.h"
@@ -178,6 +179,7 @@ static SfallOpcodeInfo opcodeInfoArray[] = {
 	{0x238, "atof",                       op_atof,                      1, true,   0, {ARG_STRING}},
 	{0x239, "scan_array",                 op_scan_array,                2, true,  -1, {ARG_OBJECT, ARG_ANY}},
 	{0x23a, "get_tile_fid",               op_get_tile_fid,              1, true,   0, {ARG_INT}},
+	{0x23b, "modified_ini",               op_modified_ini,              0, true},
 	{0x23c, "get_sfall_args",             op_get_sfall_args,            0, true},
 	{0x23d, "set_sfall_arg",              op_set_sfall_arg,             2, false,  0, {ARG_INT, ARG_ANY}}, // hook script system will validate type
 	{0x241, "get_npc_level",              op_get_npc_level,             1, true,  -1, {ARG_INTSTR}},
@@ -409,7 +411,6 @@ void Opcodes::InitNew() {
 	opcodes[0x227] = op_refresh_pc_art;
 	opcodes[0x22c] = op_stop_sfall_sound;
 
-	opcodes[0x23b] = op_modified_ini;
 	opcodes[0x23e] = op_force_aimed_shots;
 	opcodes[0x23f] = op_disable_aimed_shots;
 	opcodes[0x240] = op_mark_movie_played;
